requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
basemap==1.0.8
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
boto3==1.5.9
botocore==1.8.23
Bottleneck==1.2.1
bs4==0.0.1
cachetools==2.0.1
census==0.8.3
cffi==1.10.0
chardet==3.0.3
citipy==0.0.5
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.2
contextlib2==0.5.5
cryptography==1.8.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.4
decorator==4.0.11
dill==0.2.7.1
distributed==1.16.3
docutils==0.13.1
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
fbprophet==0.2.1
Flask==0.12.2
Flask-Cors==3.0.2
Flask-PyMongo==0.5.1
Flask-SQLAlchemy==2.3.2
future==0.16.0
gevent==1.2.1
google-api-python-client==1.6.4
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-language==0.29.0
google-gax==0.15.15
googleapis-common-protos==1.5.2
graphviz==0.8.2
greenlet==0.4.12
grpcio==1.4.0
gunicorn==19.7.1
h5py==2.7.1
HeapDict==1.0.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.5
imagesize==0.7.1
inflection==0.3.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
kdtree==0.15
Keras==2.1.3
lazy-object-proxy==1.2.2
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.0.2
menuinst==1.4.7
mistune==0.7.4
more-itertools==4.0.1
mpmath==0.19
mrjob==0.6.1
msgpack-python==0.4.8
multipledispatch==0.4.9
mysqlclient==1.3.12
navigator-updater==0.1.0
nb-conda-kernels==2.1.0
nbconvert==5.1.1
nbformat==4.3.0
ndg-httpsclient==0.4.3
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.14.0
numpydoc==0.6.0
oauth2client==4.1.2
oauthlib==2.0.3
odo==0.5.0
olefile==0.44
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
partd==0.3.8
path.py==10.3.1
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.1.1
plotly==2.2.3
ply==3.8
prompt-toolkit==1.0.14
protobuf==3.4.0
psutil==5.2.2
py==1.4.33
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pymongo==3.5.1
PyMySQL==0.7.11
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pyproj==1.9.5.1
pyreadline==2.1
pyshp==1.2.12
pystan==2.17.1.0
pytest==3.0.7
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
pywin32==220
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
Quandl==3.3.0
requests==2.14.2
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rope-py3k==0.9.4.post1
rsa==3.4.2
s3transfer==0.1.12
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.7.1
selenium==3.6.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
sphinx==1.5.6
splinter==0.7.7
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
testpath==0.3
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
tweepy==3.5.0
unicodecsv==0.14.1
uritemplate==3.0.0
vaderSentiment==2.5
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
win-unicode-console==0.5
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
